services/risk-engine/pyproject.toml
[tool.poetry]
name = "risk-engine-service"
version = "0.1.0"
description = "Fusion Prime risk engine microservice"
authors = ["Fusion Prime Team <dev@fusionprime.local>"]
readme = "README.md"
packages = [
  { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = { version = "^0.38.0", extras = ["standard"] }
pydantic = "^2.12.3"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.23"
alembic = "^1.13.1"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.9"
numpy = "^1.24.3"
pandas = "^2.1.4"
scipy = "^1.11.4"
scikit-learn = "^1.3.2"
google-cloud-bigquery = "^3.13.0"
google-cloud-pubsub = "^2.18.4"
google-cloud-logging = "^3.8.0"
google-cloud-monitoring = "^2.16.0"
redis = "^5.0.1"
aioredis = "^2.0.1"
httpx = "^0.25.2"
aiohttp = "^3.9.1"
cryptography = "^41.0.0"
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
structlog = "^23.2.0"
prometheus-client = "^0.19.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
httpx = "^0.25.2"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
